requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.14.0
notion-client>=2.2.1
python-dotenv>=1.0.0
